apps/python-semantic/requirements.txt
flask
flask-cors
pydantic
python-dotenv
openai>=1.40.0
supabase